+++ DOCS/CHANGELOG.md
@@ -33,7 +33,7 @@
     - `CrosstermService`.
     - `MiniquadPixels`, `MiniquadService`.
   - namespaces: `Fmt`, `Fs`, `FsPath`, `Io`, `Iter`, `Log`.
-- new macros: `maybe!`, `miniquad!`, `xorshift_custom!`.
+- new macros: `maybe!`, `miniquad!`, `strjoin!`, `xorshift_custom!`.
 - new modules:
   - data: `{codec::{self, radix}, list, key, table, uid, xipher}`.
   - lang: `{c, glsl}`.